genai/workshop/01_Advanced_Stock_Analysis_Using_Bedrock_Agent/requirements.txt
boto3
numpy
streamlit
PyPDF2
faiss-cpu
yfinance
plotly
